requirements.txt
discord.py
itertools
asyncio
configparser
os
embed_colors
sys
pytz
datetime
